# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: KUnnect API
  version: 1.0.0
  description: |
    KUnnect 백엔드 API 계약 (docs/dev-plan-backend.md 5절 기준, 계약 우선).
    - 인증: 세션 쿠키 `kunnect_sid` (HttpOnly). 동일 출처(Next.js rewrites 프록시) 전제, CORS 비활성.
    - ID는 정수. 시각은 ISO 8601 UTC 문자열.
    - 목록은 커서 페이지네이션: `?limit=20&cursor=<opaque>` -> `{items, nextCursor}`.
    - 오류는 항상 `{error:{code,message,fields?}}`.
    - 비로그인 열람 불가(모든 API는 로그인 필요, signup/login/health 제외).
    - 세션 쿠키가 있는 상태 변경 요청(POST/PATCH/DELETE)은 `Origin` 헤더 필수(없거나 허용 목록 밖이면 403 `BAD_ORIGIN`).
    - 공통 오류 코드(엔드포인트별로 반복 기재하지 않음): 400 `VALIDATION_ERROR`(fields 포함) / `BAD_REQUEST`(깨진 JSON 등) / `FUTURE_TIME`(fields.occurredAt) / `WEAK_PASSWORD`, 401 `UNAUTHENTICATED`, 403 `FORBIDDEN` / `BAD_ORIGIN`, 404 `NOT_FOUND`, 409 `CONFLICT` 계열(각 엔드포인트 참조: `POST_CLOSED`, `READ_ONLY`, `ALREADY_DECIDED`, `PHOTO_IN_USE` 등), 413 `PAYLOAD_TOO_LARGE`(JSON 100KB 초과) / `FILE_TOO_LARGE`, 415 `UNSUPPORTED_TYPE`, 422 `PHOTO_NOT_OWNED` / `CORRUPT_IMAGE`, 429 `RATE_LIMITED`(요청 한도, 하루 글 작성 10건, 임시 사진 20장 포함), 500 `INTERNAL`.
    - 쪽지 알림 배지: `GET /notifications` 의 `unreadCount` 는 쪽지(MESSAGE) 알림을 **포함**한 읽지 않은 알림 수이고, `/notifications/unread-count`·`/me` 의 `unread.notifications` 는 쪽지 알림을 제외한다(쪽지는 `unread.messages` 로 따로 센다).
...
    - 글 열람 범위 [가정/제안]: 작성자가 종료(CLOSED)한 글과 탈퇴한 작성자의 글은 작성자 본인 외에는 404, `?status=CLOSED` 목록은 내 글만.
    - `occurredAt` 은 오프셋이 있는 ISO 8601 문자열(예: `2026-10-01T10:00:00+09:00`, `...Z`)만 허용하며 2020-01-01 이전은 거부.
servers:
  - url: /api/v1
...
security:
  - cookieAuth: []
paths:
...
  /me/posts:
    get:
      tags: [posts]
      summary: 내가 쓴 글
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Cursor'}, {name: type, in: query, schema: {$ref: '#/components/schemas/PostType'}}, {name: status, in: query, schema: {$ref: '#/components/schemas/PostStatus'}}]
      responses:
        '200': {description: 성공, content: {application/json: {schema: {$ref: '#/components/schemas/PostCardPage'}}}}
...
  /posts:
    get:
      tags: [posts]
      summary: 글 목록 (로그인 필요)
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {name: type, in: query, schema: {$ref: '#/components/schemas/PostType'}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/PostStatus'}}
        - {name: locationId, in: query, schema: {type: integer}, description: 특정 건물·층(Location 행) 하나}
        - {name: buildingId, in: query, schema: {type: string}, description: 건물 전체(Location.buildingId, 모든 층 포함). locationId 와 함께 쓰면 둘 다 만족해야 함}
        - {name: tag, in: query, schema: {type: string}}
        - {name: q, in: query, schema: {type: string}}
      responses:
        '200': {description: 성공, content: {application/json: {schema: {$ref: '#/components/schemas/PostCardPage'}}}}
    post:
      tags: [posts]
      summary: 글 작성 (매칭은 비동기)
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/PostCreate'}}}
      responses:
        '201': {description: 생성됨 (PostDetail 평평한 객체, GET/PATCH 와 동일), content: {application/json: {schema: {$ref: '#/components/schemas/PostDetail'}}}}
        '400': {$ref: '#/components/responses/Validation'}
        '422': {description: PHOTO_NOT_OWNED, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '429': {$ref: '#/components/responses/RateLimited'}
  /posts/{id}:
    get:
      tags: [posts]
      summary: 글 상세 (hiddenFeatures는 작성자에게만)
      parameters: [{$ref: '#/components/parameters/IdPath'}]
      responses:
        '200': {description: 성공, content: {application/json: {schema: {$ref: '#/components/schemas/PostDetail'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [posts]
      summary: 글 수정 (작성자, 유형·위치 변경 불가)
      parameters: [{$ref: '#/components/parameters/IdPath'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/PostUpdate'}}}
      responses:
        '200': {description: 성공, content: {application/json: {schema: {$ref: '#/components/schemas/PostDetail'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: POST_CLOSED, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    delete:
      tags: [posts]
      summary: 글 삭제 (작성자, 즉시 영구 삭제) [가정/제안 A32]
      description: |
        작성자 본인의 글을 **즉시 영구 삭제**한다(이전에는 CLOSED 로 전환만 했음). OPEN/MATCHED/CLOSED/RETURNED 어느 상태든 가능.
        - 사진 원본·AI 사본·흐림 사본 파일, 댓글, 태그 연결, 이 글의 매칭과 알림, 인수 요청 기록이 함께 삭제된다. 이 글에서 시작된 쪽지 대화는 남고 메시지는 유지되며 글 맥락(`postContext`)만 사라진다. 신고 기록(snapshot)은 보존된다.
        - 삭제 후 `GET /posts/{id}` 와 `/me/posts`, 목록·검색에서 사라지고, 같은 글을 다시 `DELETE` 하면 **404**(이미 삭제됨; 작성자 확인을 위해 204 멱등으로 만들지 않음). 다른 사람의 글은 403.
        - 진행 중인 인수 요청(REQUESTED/VERIFIED)이 있으면 409 `ACTIVE_HANDOVER`(먼저 인수를 완료하거나 거절해야 함).
        - 조용히 닫기만 하려면 `POST /posts/{id}/status {status: CLOSED}`(종료 24시간 후 자동 삭제)를 쓴다.
      parameters: [{$ref: '#/components/parameters/IdPath'}]
      responses:
        '204': {description: 삭제됨}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {description: ACTIVE_HANDOVER (진행 중인 인수 요청이 있어 삭제할 수 없음), content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  /posts/{id}/status:
    post:
      tags: [posts]
      summary: 상태 변경 (CLOSED만 직접 가능, RETURNED는 인수 완료 플로우로만)
      parameters: [{$ref: '#/components/parameters/IdPath'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [status], properties: {status: {type: string, enum: [CLOSED]}}}}}
      responses:
        '200': {description: 성공, content: {application/json: {schema: {$ref: '#/components/schemas/PostDetail'}}}}
        '409': {description: INVALID_TRANSITION, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  /notifications:
    get:
      tags: [notifications]
      summary: 알림 목록
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Cursor'}]
      responses:
        '200': {description: 성공, content: {application/json: {schema: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Notification'}}, unreadCount: {type: integer}, nextCursor: {type: string, nullable: true}}}}}}
...
  /notifications/{id}/read:
    post:
      tags: [notifications]
      summary: 읽음 처리
      parameters: [{$ref: '#/components/parameters/IdPath'}]
      responses:
        '204': {description: 성공}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    cookieAuth: {type: apiKey, in: cookie, name: kunnect_sid}
  parameters:
    IdPath: {name: id, in: path, required: true, schema: {type: integer}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
    Cursor: {name: cursor, in: query, schema: {type: string}, description: 이전 응답의 nextCursor (불투명 값)}
...
  responses:
    Validation: {description: VALIDATION_ERROR, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    Forbidden: {description: FORBIDDEN, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    NotFound: {description: NOT_FOUND, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    RateLimited: {description: RATE_LIMITED (Retry-After 헤더), content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string, description: "VALIDATION_ERROR, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, INTERNAL 또는 엔드포인트별 코드"}
            message: {type: string, description: 한국어 메시지}
            fields: {type: object, additionalProperties: {type: string}}
    PostType: {type: string, enum: [LOST, FOUND]}
    PostStatus: {type: string, enum: [OPEN, MATCHED, RETURNED, CLOSED]}
...
    AuthorRef:
      type: object
      properties: {id: {type: integer}, nickname: {type: string}}
    Location:
      type: object
      description: "목록(/locations)은 표시 순서대로 내려간다. buildingId 'etc'(\"기타\")는 글의 locationText 로 장소를 직접 적는 항목이다."
      properties: {id: {type: integer}, buildingId: {type: string, description: 같은 건물의 모든 층이 공유하는 식별자}, buildingName: {type: string}, floor: {type: integer, nullable: true}, lat: {type: number}, lng: {type: number}}
...
    Photo:
      type: object
      description: |
        요청자 관점의 사진 정보 [가정/제안 — 민감 사진 흐림].
        민감 사진(학생증·카드·얼굴 등)은 작성자와 소유 확인(VERIFIED) 이후의 인수 상대에게만 원본(`isBlurred=false`)이 내려가고,
        그 외 사용자에게는 흐림 처리된 사본 URL(`isBlurred=true`)이 내려간다. 흐림 사본을 만들 수 없으면 `url=null`(원본으로 대체하지 않음).
        `sensitive`: 작성자에게는 유효 민감 플래그, 그 외에는 `isBlurred` 와 같다.
      properties:
        photoId: {type: integer}
        url: {type: string, nullable: true}
        width: {type: integer}
        height: {type: integer}
        isBlurred: {type: boolean}
        sensitive: {type: boolean}
...
    PostCard:
      type: object
      properties:
        id: {type: integer}
        type: {$ref: '#/components/schemas/PostType'}
        title: {type: string}
        status: {$ref: '#/components/schemas/PostStatus'}
        thumbnailUrl: {type: string, nullable: true, description: '민감 사진은 작성자 외에는 흐림 사본 URL (사본 없으면 null)'}
        hasSensitivePhoto: {type: boolean}
        locationName: {type: string}
        locationText: {type: string, nullable: true, description: "위치가 기타(etc)일 때 작성자가 입력한 장소. 그 외 null. 표시는 클라이언트가 조합(예: 기타 · {locationText})"}
        occurredAt: {type: string, format: date-time}
        tags: {type: array, items: {type: string}}
        author: {$ref: '#/components/schemas/AuthorRef'}
        createdAt: {type: string, format: date-time}
    PostCardPage:
      type: object
      properties: {items: {type: array, items: {$ref: '#/components/schemas/PostCard'}}, nextCursor: {type: string, nullable: true}}
    PostDetail:
      type: object
      properties:
        id: {type: integer}
        type: {$ref: '#/components/schemas/PostType'}
        title: {type: string}
        description: {type: string}
        status: {$ref: '#/components/schemas/PostStatus'}
        matchState: {type: string, enum: [PENDING, DONE, FAILED]}
        occurredAt: {type: string, format: date-time}
        location: {$ref: '#/components/schemas/Location'}
        locationText: {type: string, nullable: true, description: 위치가 기타(etc)일 때 입력한 장소. 그 외 null}
        lat: {type: number, nullable: true}
        lng: {type: number, nullable: true}
        storagePlace: {type: string, nullable: true}
        hiddenFeatures: {type: string, nullable: true, description: 작성자 본인에게만 노출}
        photos: {type: array, items: {$ref: '#/components/schemas/Photo'}}
        hasSensitivePhoto: {type: boolean}
        tags: {type: array, items: {type: string}}
        author: {$ref: '#/components/schemas/AuthorRef'}
        commentCount: {type: integer}
        isMine: {type: boolean}
        createdAt: {type: string, format: date-time}
    PostCreate:
      type: object
      required: [type, title, description, locationId, occurredAt]
      properties:
        type: {$ref: '#/components/schemas/PostType'}
        title: {type: string, maxLength: 50}
        description: {type: string, maxLength: 1000}
        locationId: {type: integer}
        lat: {type: number}
        lng: {type: number}
        occurredAt: {type: string, format: date-time}
        tags: {type: array, maxItems: 8, items: {type: string, maxLength: 20}}
        storagePlace: {type: string, maxLength: 100, description: FOUND 필수}
        hiddenFeatures: {type: string, maxLength: 300}
        locationText: {type: string, nullable: true, maxLength: 50, description: "위치가 기타(etc)면 필수(앞뒤 공백 제거 후 1~50자), 그 외 위치는 비우거나 null(값이 있으면 400 VALIDATION_ERROR)"}
        photoIds: {type: array, maxItems: 3, items: {type: integer}}
    PostUpdate:
      type: object
      properties:
        title: {type: string, maxLength: 50}
        description: {type: string, maxLength: 1000}
        tags: {type: array, maxItems: 8, items: {type: string, maxLength: 20}}
        storagePlace: {type: string, maxLength: 100}
        hiddenFeatures: {type: string, maxLength: 300}
        locationText: {type: string, nullable: true, maxLength: 50, description: "위치가 기타(etc)인 글만 수정 가능(1~50자). 그 외 위치에서 값을 보내면 400"}
...
    Notification:
      type: object
      properties:
        id: {type: integer}
        type: {type: string, enum: [MATCH, COMMENT, REPLY, MESSAGE]}
        text: {type: string, description: 서버 생성 문구 (물건·위치 비포함)}
        createdAt: {type: string, format: date-time}
        readAt: {type: string, format: date-time, nullable: true}
        target:
          type: object
          properties: {kind: {type: string, enum: [match, comment, conversation]}, id: {type: integer}, postId: {type: integer, nullable: true}}
...